refactor: Reorganize archive service file
