+++ node/src/transition_frontier/archive/archive_config.rs
@@ -1,5 +1,5 @@
 use serde::{Deserialize, Serialize};
-use std::path::PathBuf;
+// use std::path::PathBuf;
 
 // TODO(adonagy): Do we need this? Is it just unnecessary boilerplate?
